Update patch set 3
Patch Set 3:

> Patch Set 1:
> 
> Could you upload the original CL first, then add your changes as a separate patchset so I can review the diff? Also, did you figure out what was causing the chromium tests to break?

Patch2 is the original SrtpTransport CL + fixing the session recreation.

Patch3 fixs the Chromium tests. The issue is that we can only enable the external authentication when creating a new SrtpSession but we did that for existing session as well which caused some problems. I verified that locally but the trybots are red because of patch failure and I guess rolling WebRTC to Chromium could fix that.

Patch-set: 3
Reviewer: Zhi Huang <5080@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0