Update patch set 8
Patch Set 8:

Android failure is due to the macro 

#define CHECK_RELEASE(ptr) \
  RTC_CHECK_EQ(0, (ptr)->Release()) << "Unexpected refcount."

It can't be changed to use HasOneRef, because that's not in RefCountInterface.
It would make some sense to me to let Release() have a bool return value, true if the object was deleted.

Patch-set: 8
Reviewer: Niels Moller <5234@58829da1-049c-39fb-b951-ebdcd0984767>