WIP rtc::AtomicOps: Add atomic ReleaseStorePtr
This CL adds atomic operations for storing a pointer.

I'm porting some Chromium JNI code that I want to use, and I'm replacing the Chromium atomics with WebRTC atomics. This CL adds an atomic operation that is lacking in the WebRTC code. I'm not sure I'm doing it correctly and I have just pattern matched the existing code. Especially the Windows code is worrying, because I have no idea if what I'm doing is correct.

On a side note, <atomic> is now an allowed C++11 library feature, and Chromium is already using it. Would it be better to start using that in WebRTC instead?

Bug: webrtc:8278
Change-Id: Icd57beae0ae94d5267df0ec79326f4668fe2d719