Update patch set 3
Patch Set 3:

My main concern is that the client now has to have two references instead of a single one. This feels like adding unnecessary complexity to the client.

Is the idea is that the client might want to use a different SurfaceView that implements different kind of layout logic? Maybe it would make sense to have the SurfaceVideoView directly use SurfaceVideoRenderer for rendering? This should hopefully be easy. Also, I think it makes sense for SurfaceVideoRenderer to extend EglRenderer.

Patch-set: 3
Reviewer: Sami Kalliomäki <5112@58829da1-049c-39fb-b951-ebdcd0984767>