Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> My main concern is that the client now has to have two references instead of a single one. This feels like adding unnecessary complexity to the client.
> 
> Is the idea is that the client might want to use a different SurfaceView that implements different kind of layout logic? Maybe it would make sense to have the SurfaceVideoView directly use SurfaceVideoRenderer for rendering? This should hopefully be easy. Also, I think it makes sense for SurfaceVideoRenderer to extend EglRenderer.

Not exactly. We do use TextureView for animation on pre-Lollipop systems but the main goal is to remove the dependency on the view. We run everything related to MediaCodec in a separate service process to deal with buggy systems and the renderer only has access to a gui supplied Surface.

I will update the CL to address your concern. But I think most real world use cases would not want a session killed along with the activity.

Patch-set: 3
Reviewer: Xiaolei Yu <5510@58829da1-049c-39fb-b951-ebdcd0984767>