Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 4: Code-Review-1
> 
> (1 comment)
> 
> Not lgtm; removing the unnecessary references should be done in a separate CL as one that changes how ownership is transferred. Which is a topic that's currently under discussion in the webrtc core team.

It's not possible to split it up, they are related. The video factories are passed through CreateModularPeerConnectionFactory() and then to PeerConnectionFactory which keeps ownership. This does not change the ownership model on the level of the CreatePeerConnectionFactory() API. I agree it's ugly they are passed in as raw pointers when we take ownership, but that's an issue outside the scope of this CL. I have to convert them from raw pointers to std::unique_ptrs at some point. Today it happens in the PeerConnectionFactory.

Can you give a link to the discussion with the core team about this, I would like to be part of it or at least follow what you say.

Patch-set: 4
Reviewer: Magnus Jedvert <5098@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0