Update patch set 7
Patch Set 7: Code-Review+1

(2 comments)

Awesome! Great!

One more change should be done as well though, as you still are using an Optional as input to a method that is never called with empty Optionals.

Please check the comments in the code.

LGTM conditioned on that that is addressed.

Patch-set: 7
Reviewer: Per Åhgren <5125@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1