Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (10 comments)
> 
> Nice CL! In general, add a unit to the variable names (s or ms). Also, you can reduce code duplication a lot in the test.

Thanks for good comments!

I completely agree that I should get rid of the code duplication in the test.

Units are nice, but it would make the variable names less similar to the names in the standard. I guess it is worth it though.

Patch-set: 1
Reviewer: Gustaf Ullberg <5116@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0