Update patch set 1
Patch Set 1:

Where is this chromium test runner coming from? Is it inherently provided by the java test binaries or do you have to depend on anything in the Chromium source code? The reason I'm asking is that we have some test code that depends directly on Chromium code (base/), which is deeply unhealthy, and we don't want more of that.

Patch-set: 1
CC: Patrik Höglund <5027@58829da1-049c-39fb-b951-ebdcd0984767>