Update patch set 4
Patch Set 4:

I think you have extended the scope of this way past what I was thinking. The important bit is maybe not the exact pattern that is use by the encoder, but only the verification that this won't lead to corruptions be referencing buffers with data from higher layers and making sure the sync bit is only set when it is valid.

The screenshare_layers is especially vulnerable since the allocation there is dynamic rather than hard-coded, and so should be prioritized.

Could you make a generic verifier that applicable to all temporal layers implementations, and if you'd like to still keep the exact validation you did for the default layer, then that can just extend that generic one?

Patch-set: 4
Reviewer: Erik Språng <5524@58829da1-049c-39fb-b951-ebdcd0984767>