Update patch set 1
Patch Set 1:

Failing test is RtpFecTest/1.FecRecoveryWithSeqNumGapTwoFrames, describe like this:

  // Two frames: first frame (old) with two media packets and 1 FEC packet.
  // Second frame (new) with 3 media packets, and no FEC packets.
  //       ---Frame 1----                     ----Frame 2------
  //  #0(media) #1(media) #2(FEC)     #65535(media) #0(media) #1(media).
  // If we lose either packet 0 or 1 of second frame, FEC decoding should not
  // try to decode using "old" FEC packet #2.

Not sure how to think about this. I'm consider trying to run with test but with sequence number shifted to not involve any wraparound, to see if that fails before or after this change. If that fails before my cl, I'd argue the test is incorrect.

Patch-set: 1
Reviewer: Niels Moller <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false