Introduce RtcEvent and subclasses
We're moving to an RtcEventLog interface that accepts std::unique_ptr<EventLog> and stores the event for encoding when encoding becomes necessary, rather than before. This will be useful while we maintain the legacy (current) encoding alongside the new encoding on which we're working.

This CL just introduces the new RtcEvent and its sub-classes, without constructors and without use. Upcoming CLs will finish the work.

BUG=webrtc:8111

Change-Id: I782383e861c31670b7cd13ffc6b43ca6a26c98f6