Add GetRemoteAudioSSLCertificate() to the PeerConnection API
This method allows the client to get details about the SSL
certificate sent by the remote side in the DTLS handshake.

This functionality in this new method has been standardized in the
RTCDtlsTransport, but until we have that implemented we wish to
expose this functionality in the public interface so clients do
not need to depend on WebRtcSession.

Bug: webrtc:8323
Change-Id: Ic964266dd7e734cec07289a147fd8d090d74ce6b