Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What are the changes vs the previous attempt?

None, so far. I'm not sure how relands are supposed to be done with gerrit, I just pressed the shiny "Reland" button on the original cl.

The interesting part of the cl is the single line change to  rtc_base/refcount.h, deleting an #include line. That's what appeared to break fyi bots post-commit, and that is unchanged in the reland.

So I'm a bit puzzled.

Patch-set: 1
Reviewer: Niels Moller <5234@58829da1-049c-39fb-b951-ebdcd0984767>