Fixing invalid calls to FindMatchingCodec.
The first argument of FindMatchingCodec is supposed to be the list that
contains the codec to be found, specifically to handle RTX codecs that
point to other codecs. But this wasn't being done everywhere, and wasn't
noticed because *most of the time* it just results in adding the RTX
codec in a different location, which isn't an issue.

But, it's still not standards-compliant. And it sometimes is an issue
when talking to older endpoints.

Adding a regression test, and DCHECK in FindMatchingCodec to ensure this
doesn't happen by accident again.

Bug: webrtc:8332
Change-Id: I5def056b245c6d00a49a59d429f1dee303fb7cef