Add support for winuwp to examples/webrtc_unity_plugin 🎉
This adds support for `winuwp` to `examples/webrtc_unity_plugin` behind the `WINUWP` build flag.
Note that the only way to build this for `winuwp` at this time is to use the vs wrapper project, as defined in https://github.com/webrtc-uwp/webrtc-windows/pull/48
this is because the winuwp wrapper layer (from webrtc-uwp project) does not yet support GN build targets.