Fix invalid IAudioSessionControl state DCHECK
See https://bugs.chromium.org/p/webrtc/issues/detail?id=11833 for
details. This change comments out a DCHECK assertion which assumes the
state of the audio session control to be inactive just after creating
the audio client, which is incorrect since the audio session control
lifetime is tied to the audio session and not the audio client, so can
stay active after the client is destroyed and re-created.